Refactored the repo
