--- requirements.txt
@@ -1,6 +1,0 @@
-numpy>=1.19.4
-scipy>=1.4.1
-opencv-python-headless
-scikit-image>=0.19.3
-matplotlib>=3.6.3
-Pillow>=9.4.0
